refactor: remove deprecated threads API endpoints and update threadsService interface
- Deleted the threads API endpoints for creating and retrieving threads, as they are no longer needed.
- Updated the threadsService interface to include method signatures for thread operations, ensuring type safety and clarity in service usage.
