+++ src/api/api-v1/paths/threads/index.ts
@@ -1,10 +1,42 @@
 // ./api/api-v1/paths/threads.ts
 
 import { Router } from "express";
-import threadsService from "@/api/api-v1/services/threadsService";
+import { ThreadsService } from "@/api/api-v1/services/threadsService";
 
-export default function (service: typeof threadsService) {
+export default function (threadsService: ThreadsService) {
     const router = Router();
+
+    /**
+     * @swagger
+     * /threads/{id}:
+     *   get:
+     *     summary: Get modeling thread in MINT.
+     *     operationId: getThread
+     *     tags: [Threads]
+     *     security:
+     *       - BearerAuth: []
+     *         oauth2: []
+     *     parameters:
+     *       - in: path
+     *         name: id
+     *         required: true
+     *         schema:
+     *           type: string
+     *     responses:
+     *       200:
+     *         description: Successful response
+     *       default:
+     *         description: An error occurred
+     */
+    router.get("/:id", async (req, res) => {
+        try {
+            const response = await threadsService.getThread(req.params.id);
+            if (response === undefined) res.status(404).send();
+            else res.status(200).send(response);
+        } catch (error) {
+            res.status(500).send(error);
+        }
+    });
 
     /**
      * @swagger
@@ -31,7 +63,7 @@ export default function (service: typeof threadsService) {
      */
     router.post("/", async (req, res) => {
         try {
-            const result = await service.createThread(req.body);
+            const result = await threadsService.createThread(req.body);
             if (result.result === "error") {
                 res.status(406).json(result);
             } else {
